Perfected _Main mechanism so _Main can be dragged into Hierarchy
Perfected _Main mechanism so _Main can be dragged into Hierarchy, for instance because we want to access the child objects parented under _Main.

This new version of _Main prevents spawning a second copy of _Main if one is already present in the Hierarchy and the one in the Resources has been marked as inactive.

NOTE: When we instantiate _Main from Resources, we change its name from _Main(Clone) to simply _Main, so that there is no difference in its name whether the _Main was instantiated from Resources or dragged from Resources into the Hierarchy manually.